+++ src/orchestration/cli/runner.js
@@ -20,8 +20,8 @@ import { runHookMessage } from '../commit-message/runner.js';
 import { runInit, runInstallHooks } from '../../commands/init.js';
 import { runInstallCiCommand } from './install-ci.js';
 import { runPrePush } from '../../commands/pre-push.js';
-import { runPreCommit } from '../../commands/pre-commit.js';
 import { runQualityFileCommand } from '../pre-commit/quality-command.js';
+import { runPreCommit } from '../pre-commit/runner.js';
 
 const registeredManualGates = gateRegistry.all
   .filter(({ manualCommand }) => manualCommand)